Refactor profile-gateway-core to profile-tenant-core
